Fully ported VectorWars to C# and removing the C++ files, though C# version has bugs - Now in the process of getting the ported files to behave the same as the original C++

+++ src/apps/VectorWarSharp/DataStructure/PlayerConnectionInfo.cs
@@ -9,6 +9,6 @@ namespace VectorWar.DataStructure
         public PlayerConnectState state;
         public int connectProgress;
         public int disconnectTimeout;
-        public int disconnectStart;
+        public long disconnectStart;
     }
 }
